Model dicts in Neo4J as lists
In Neo4J it is not possible to save dicts or list of dicts in a node property.
By creating separate nodes for each dict and connecting them with relationships, we complicate the model and the Neo4J writing process will be longer.
That's why we decided to model lists of dicts as multiple lists where each list represent values saved under one key.