Handle props modeled differently in Neo4J than in Metamodel
Some properties are modeled in Neo4J not like in AAS Metamodel.
E.g. some properties which are List of dicts are modelled as
multiple list properties to reduce num of nodes and rels.

This commit adjust `_convert_referable_subgraph_to_dict` such
that it can convert a Referable subgraph back to metamodel
conform JSON